Revert "A_Star改改"
This reverts commit ba3ce3fdfad753f092c7a2cd41a2510551da2ebc.

+++ Src/Astar/A_STAR.h
@@ -6,7 +6,7 @@
 
 typedef struct {
 	int loc[2];
-	double data[3];   //G H F
+	float data[3];   //G H F
 	int state;
 	int occupy;
 	int position;
